Add example for "Visual Camera"

+++ OpenGL 4 Tutorials/Vertex Buffer Objects/glutils.jl
@@ -139,7 +139,7 @@ function startgl()
     version = bytestring(glGetString(GL_VERSION))
     println("Renderder: ", renderer)
     println("OpenGL version supported: ", version)
-    @assert parse(Float32, version[1:3]) >= 3.2 "OpenGL version must ≥ 3.2, Please upgrade your graphic driver."
+    @assert parse(Float64, version[1:3]) >= 3.2 "OpenGL version must ≥ 3.2, Please upgrade your graphic driver."
     # save logs
     logadd("gl.log", string("renderer: ", renderer, "\nversion: ", version))
     glparams()
